fix: fix check style and remove code duplication
